#1112 audit log 大容量対応

+++ platform_root/platform_job/jobs/AuditLogJobExecutor.py
@@ -38,7 +38,9 @@ from libs import queries_auditlog
 
 TEMP_FILENAME_PREFIX = "exastro_auditlog_"
 TEMPORARY_DIR = os.environ.get('TEMPORARY_DIR')
-BUFFER_SIZE=1024000
+BUFFER_SIZE=10_240_000  # 10MB
+GET_ROWS_LIMIT=1000
+
 
 class AuditLogJobExecutor(BaseJobExecutor):
     """監査ログファイル生成Job / Audit log file generation job
@@ -96,59 +98,96 @@ class AuditLogJobExecutor(BaseJobExecutor):
                 # 抽出条件
                 conditions = json.loads(t_jobs_audit_log["CONDITIONS"])
 
+                # T_AUDIT_LOG 取得範囲・件数取得
+                #   データ取得の際にIDのRANGEで分割取得するため（RANGEなしで１回で大量データ取得するとDBのCPU高騰するための対策）
+                globals.logger.debug('Start Get T_AUDIT_LOG Summary')
+                with conn.cursor() as cursor:
+                    cursor.execute(queries_auditlog.SQL_QUERY_AUDIT_LOG_SUMMARY, conditions)
+                    t_audit_log_summary = cursor.fetchone()
+
+                globals.logger.debug(f'Finish Get T_AUDIT_LOG Summary: MIN_ID={t_audit_log_summary["MIN_ID"]} / MAX_ID={t_audit_log_summary["MAX_ID"]} / COUNT={t_audit_log_summary["COUNT_ROWS"]}')
+                if t_audit_log_summary["COUNT_ROWS"] == 0:
+                    # 対象データがない場合
+                    self.__update_job_audit_log(conn, job_status=const.AUDIT_LOG_NODATA, message=None)
+                    return True
+
                 with conn.cursor() as cursor, \
-                    tempfile.NamedTemporaryFile(mode='w', encoding='utf-8', dir=TEMPORARY_DIR, prefix=TEMP_FILENAME_PREFIX, suffix='.log', delete=False) as fp_temp, \
                     tempfile.NamedTemporaryFile(dir=TEMPORARY_DIR, prefix=TEMP_FILENAME_PREFIX, suffix='.zip', delete=False) as fp_zip:
-
                     try:
-                        # SELECT T_AUDIT_LOG
-                        globals.logger.debug(f'Start Write Temprary File: file={os.path.join(TEMPORARY_DIR, fp_temp.name)}');
-                        cursor.execute(queries_auditlog.SQL_QUERY_AUDIT_LOG, conditions)
-                        while True:
-                            t_audit_log = cursor.fetchone()
-                            if t_audit_log is None:
-                                break
+                        # ZIPファイルopen
+                        with zipfile.ZipFile(fp_zip, mode='w', compression=zipfile.ZIP_DEFLATED) as zipfl, \
+                            zipfl.open(self.__arcname(conditions), 'w') as fp_log:
 
-                            # count_export increment
-                            self.count_export = (self.count_export + 1) if self.count_export is not None else 1
+                            globals.logger.debug(f'Start Write Zip File: file={os.path.join(TEMPORARY_DIR, fp_zip.name)}')
+                            id_from = t_audit_log_summary['MIN_ID']
 
-                            # 出力レコードの生成 / Generating output records
-                            json_audit_log = {}
-                            for key, value in t_audit_log.items():
-                                if key in ['ID','CREATE_TIMESTAMP', 'CREATE_USER', 'LAST_UPDATE_TIMESTAMP', 'LAST_UPDATE_USER']:
-                                    # 出力しない項目
-                                    pass
-                                elif isinstance(value, datetime.datetime):
-                                    json_audit_log[key.lower()] = common.datetime_to_str(value)
-                                else:
-                                    json_audit_log[key.lower()] = value
+                            while True:
+                                # 一度に取得するとDBのCPU高騰するので分割して取得する
+                                conditions['id_from'] = id_from
+                                conditions['id_to'] = id_from + GET_ROWS_LIMIT - 1
 
-                            # Write Temporary File
-                            fp_temp.write(json.dumps(json_audit_log)+'\n')
+                                # SELECT T_AUDIT_LOG
+                                cursor.execute(queries_auditlog.SQL_QUERY_AUDIT_LOG, conditions)
+                                while True:
+                                    t_audit_log = cursor.fetchone()
+                                    if t_audit_log is None:
+                                        break
 
-                            if self.count_export % job_manager_config.JOBS[const.PROCESS_KIND_AUDIT_LOG]["extra_config"]["status_update_interval"] == 0:
-                                # 件数更新のためステータス更新を行う / Update the status to update the number of items
-                                self.__update_job_audit_log(conn, job_status=const.AUDIT_LOG_EXEC, message=None)
-                                conn.commit()
+                                    # count_export increment
+                                    self.count_export = (self.count_export + 1) if self.count_export is not None else 1
 
-                            # 1JOBでリソースを占有しないようにsleepする / Sleep so that 1JOB does not occupy resources
-                            time.sleep(job_manager_config.JOBS[const.PROCESS_KIND_AUDIT_LOG]["extra_config"]["output_interval_millisecond"]/1000)
+                                    # 出力レコードの生成 / Generating output records
+                                    json_audit_log = {}
+                                    for key, value in t_audit_log.items():
+                                        if key in ['ID','CREATE_TIMESTAMP', 'CREATE_USER', 'LAST_UPDATE_TIMESTAMP', 'LAST_UPDATE_USER']:
+                                            # 出力しない項目
+                                            pass
+                                        elif isinstance(value, datetime.datetime):
+                                            json_audit_log[key.lower()] = common.datetime_to_str(value)
+                                        else:
+                                            json_audit_log[key.lower()] = value
 
-                        fp_temp.close()
-                        globals.logger.debug(f'Finish Write Temprary File: count_export={self.count_export}')
+                                    # Write Temporary File
+                                    fp_log.write((json.dumps(json_audit_log)+'\n').encode())
+
+                                    if self.count_export % job_manager_config.JOBS[const.PROCESS_KIND_AUDIT_LOG]["extra_config"]["status_update_interval"] == 0:
+                                        # 件数更新のためステータス更新を行う / Update the status to update the number of items
+                                        globals.logger.debug(f'Writing Zip File: exported={self.count_export} ({round(self.count_export / t_audit_log_summary["COUNT_ROWS"] * 100)}%)')
+                                        self.__update_job_audit_log(conn, job_status=const.AUDIT_LOG_EXEC, message=None)
+                                        conn.commit()
+
+                                id_from += GET_ROWS_LIMIT
+                                if conditions['id_to'] > t_audit_log_summary['MAX_ID']:
+                                    # MAXのIDを超えたら終了する
+                                    break
+
+                                time.sleep(job_manager_config.JOBS[const.PROCESS_KIND_AUDIT_LOG]["extra_config"]["output_interval_millisecond"]/1000)
+
+
+                        file_size = os.path.getsize(os.path.join(TEMPORARY_DIR, fp_zip.name))
+                        globals.logger.debug(f'Finish Write Zip File: count_export={self.count_export} / size={file_size}')
+
+                        # ファイルサイズ確認
+                        if file_size > job_manager_config.JOBS[const.PROCESS_KIND_AUDIT_LOG]["extra_config"]["max_file_size"]:
+                            # ファイルが大きすぎる場合、エラーとする
+                            globals.logger.info(f'File size over file_size={file_size} / max_file_size={job_manager_config.JOBS[const.PROCESS_KIND_AUDIT_LOG]["extra_config"]["max_file_size"]}')
+                            self.__update_job_audit_log(
+                                conn,
+                                job_status=const.AUDIT_LOG_FAILED,
+                                message=multi_lang.get_text_spec(
+                                    self.language,
+                                    '401-00024',
+                                    'ファイルサイズが大きすぎます。抽出条件を見直して再実行してください。(ファイルサイズ:{0}byte / 最大サイズ:{1}byte)',
+                                    file_size,
+                                    job_manager_config.JOBS[const.PROCESS_KIND_AUDIT_LOG]["extra_config"]["max_file_size"]
+                                    )
+                            )
+                            return False
 
                         if self.count_export is not None and self.count_export > 0:
                             # データがある場合のみファイル生成を行う / Generate file only if data exists
 
-                            # ZIPファイルの生成 / Generate ZIP file
-                            globals.logger.debug(f'Start Write Zip File: file={os.path.join(TEMPORARY_DIR, fp_zip.name)}')
-                            with zipfile.ZipFile(fp_zip, mode='w') as zipfl:
-                                zipfl.write(os.path.join(TEMPORARY_DIR, fp_temp.name), arcname=self.__arcname(conditions))
-
-                            globals.logger.debug(f'Finish Write Zip File');
-
-                            # INSERT T_JOBS_AUDIT_LOG_FILE
-                            globals.logger.debug(f'Insert T_JOBS_AUDIT_LOG_FILE: file_id={self.file_id}')
+                            globals.logger.debug(f'Start Write T_JOBS_AUDIT_LOG_FILE')
                             cursor.execute(queries_auditlog.SQL_INSERT_JOBS_AUDIT_LOG_FILE,{
                                 "file_id": self.file_id,
                                 "job_id": self.job_id,
@@ -157,6 +196,7 @@ class AuditLogJobExecutor(BaseJobExecutor):
                             })
 
                             # write zip file image
+                            written_size = 0
                             globals.logger.debug(f'Start Write T_JOBS_AUDIT_LOG_FILE.file_data')
                             with open(os.path.join(TEMPORARY_DIR, fp_zip.name), 'rb') as fp_zipr:
                                 while True:
@@ -168,16 +208,20 @@ class AuditLogJobExecutor(BaseJobExecutor):
                                         "file_id": self.file_id,
                                         "file_data": zip_buf
                                     })
+                                    self.__update_job_audit_log(conn, job_status=const.AUDIT_LOG_EXEC, message=None)
+                                    conn.commit()
+                                    written_size += len(zip_buf)
+                                    globals.logger.debug(f'Writing T_JOBS_AUDIT_LOG_FILE.file_data: written={written_size} ({round(written_size / file_size * 100)}%)')
 
                                     # 1JOBでリソースを占有しないようにsleepする / Sleep so that 1JOB does not occupy resources
                                     time.sleep(job_manager_config.JOBS[const.PROCESS_KIND_AUDIT_LOG]["extra_config"]["output_interval_millisecond"]/1000)
 
-                            globals.logger.debug(f'Finish Write T_JOBS_AUDIT_LOG_FILE.file_data');
+                            globals.logger.debug(f'Finish Write T_JOBS_AUDIT_LOG_FILE')
                     finally:
                         try:
                             # remove temporary files
-                            os.remove(os.path.join(TEMPORARY_DIR, fp_temp.name))
-                            os.remove(os.path.join(TEMPORARY_DIR, fp_zip.name))
+                            if os.path.isfile(os.path.join(TEMPORARY_DIR, fp_zip.name)):
+                                os.remove(os.path.join(TEMPORARY_DIR, fp_zip.name))
                         except Exception:
                             pass
 
@@ -186,23 +230,23 @@ class AuditLogJobExecutor(BaseJobExecutor):
                     self.__update_job_audit_log(conn, job_status=const.AUDIT_LOG_NODATA, message=None)
                 else:
                     # 最終的なステータスに更新する / Update to final status
-                    self.__update_job_audit_log(conn, job_status=const.JOB_USER_COMP, message=None)
+                    self.__update_job_audit_log(conn, job_status=const.AUDIT_LOG_COMP, message=None)
 
             except JobTimeoutException as ex:
                 # タイムアウトエラー / timeout error
                 conn.rollback()
-                self.__update_job_audit_log(conn, job_status=const.JOB_USER_FAILED, message=multi_lang.get_text_spec(self.language, '401-00022', 'ファイル生成中にタイムアウトしました。'))
+                self.__update_job_audit_log(conn, job_status=const.AUDIT_LOG_FAILED, message=multi_lang.get_text_spec(self.language, '401-00022', 'ファイル生成中にタイムアウトしました。'))
                 return False
 
             except common.InternalErrorException as ex:
                 globals.logger.error(f'{ex.message}')
                 conn.rollback()
-                self.__update_job_audit_log(conn, job_status=const.JOB_USER_FAILED, message=ex.message)
+                self.__update_job_audit_log(conn, job_status=const.AUDIT_LOG_FAILED, message=ex.message)
                 raise ex
 
             except Exception as ex:
                 globals.logger.error(f'{ex}\n-- stack trace --\n{traceback.format_exc()}')
-                self.__update_job_audit_log(conn, job_status=const.JOB_USER_FAILED, message=multi_lang.get_text_spec(self.language, '401-00023', 'ファイル生成中に想定外のエラーが発生し、処理を中断しました。'))
+                self.__update_job_audit_log(conn, job_status=const.AUDIT_LOG_FAILED, message=multi_lang.get_text_spec(self.language, '401-00023', 'ファイル生成中に想定外のエラーが発生し、処理を中断しました。'))
                 raise ex
 
             finally:
@@ -249,7 +293,7 @@ class AuditLogJobExecutor(BaseJobExecutor):
         try:
             globals.logger.info(f'Cancel JOB [JOB_ID:{self.job_id}]')
             with closing(DBconnector().connect_orgdb(self.organization_id)) as conn:
-                self.__update_t_jobs_user(conn, job_status=const.JOB_USER_FAILED, message=multi_lang.get_text_spec(self.language, '401-00011', '{0}行目の処理中にタイムアウトしました。', self.imp_wb.get_row_idx()))
+                self.__update_job_audit_log(conn, job_status=const.AUDIT_LOG_FAILED, message=multi_lang.get_text_spec(self.language, '401-00022', 'ファイル生成中にタイムアウトしました。'))
                 conn.commit()
             return True
         except Exception as err:
@@ -289,8 +333,12 @@ class AuditLogJobExecutor(BaseJobExecutor):
                     try:
                         with closing(DBconnector().connect_orgdb(organization['ORGANIZATION_ID'])) as conn, conn.cursor() as cursor:
 
+                            # T_JOBS_AUDIT_LOG_FILE出力中は更新日時を更新できないので、TIMEOUT時間も含める
                             last_update_timestamp = (datetime.datetime.now() - datetime.timedelta(
-                                seconds=job_manager_config.JOBS[job_manager_const.PROCESS_KIND_FORCE_UPDATE_STATUS]['extra_config']['prograss_seconds']))
+                                seconds=
+                                    job_manager_config.JOBS[job_manager_const.PROCESS_KIND_FORCE_UPDATE_STATUS]['extra_config']['prograss_seconds'] +
+                                    job_manager_config.JOBS[const.PROCESS_KIND_AUDIT_LOG]['timeout_seconds']
+                            ))
 
                             # 未完了状態で一定時間経過したものを対象とする / Targets items that have been incomplete for a certain period of time
                             cursor.execute(
